@@ -29,66 +29,68 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.0.4",
-    "@angular/common": "^6.0.4",
-    "@angular/compiler": "^6.0.4",
-    "@angular/core": "^6.0.4",
-    "@angular/forms": "^6.0.4",
-    "@angular/http": "^6.0.4",
-    "@angular/platform-browser": "^6.0.4",
-    "@angular/platform-browser-dynamic": "^6.0.4",
-    "@angular/pwa": "^0.13.3",
-    "@angular/router": "^6.0.4",
-    "@angular/service-worker": "^6.0.4",
+    "@angular/animations": "^7.2.9",
+    "@angular/cdk": "^7.3.4",
+    "@angular/common": "^7.2.9",
+    "@angular/compiler": "^7.2.9",
+    "@angular/core": "^7.2.9",
+    "@angular/fire": "^5.1.2",
+    "@angular/forms": "^7.2.9",
+    "@angular/http": "^7.2.9",
+    "@angular/platform-browser": "^7.2.9",
+    "@angular/platform-browser-dynamic": "^7.2.9",
+    "@angular/pwa": "^0.13.6",
+    "@angular/router": "^7.2.9",
+    "@angular/service-worker": "^7.2.9",
     "@fortawesome/angular-fontawesome": "0.1.0-10",
     "@fortawesome/fontawesome": "^1.1.7",
     "@fortawesome/fontawesome-svg-core": "^1.2.0-11",
     "@fortawesome/free-brands-svg-icons": "^5.1.0-8",
     "@fortawesome/free-regular-svg-icons": "^5.1.0-8",
     "@fortawesome/free-solid-svg-icons": "^5.1.0-8",
-    "@ngx-share/button": "^5.0.1",
-    "@ngx-share/buttons": "^5.0.1",
-    "@ngx-share/core": "^5.0.1",
-    "angular-bootstrap-md": "^6.0.1",
+    "@ngx-share/button": "^7.1.2",
+    "@ngx-share/buttons": "^7.1.2",
+    "@ngx-share/core": "^7.1.2",
+    "acorn": "^6.1.1",
+    "angular-bootstrap-md": "^7.4.2",
     "angular-font-awesome": "^3.1.2",
-    "angularfire2": "^5.0.0-rc.7",
     "body-parser": "~1.18.3",
     "core-js": "^2.4.1",
     "crawler-user-agents": "git+https://github.com/monperrus/crawler-user-agents.git",
     "express": "~4.16.3",
-    "firebase": "5.0.4",
+    "firebase": "5.9.0",
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
     "method-override": "~2.3.10",
-    "rxjs": "^6.2.0",
+    "rxjs": "^6.4.0",
     "rxjs-compat": "^6.2.0",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.8",
-    "@angular/cli": "^6.0.8",
-    "@angular/compiler-cli": "^6.0.4",
-    "@angular/language-service": "^6.0.4",
+    "@angular-devkit/build-angular": "^0.13.6",
+    "@angular/cli": "^7.3.6",
+    "@angular/compiler-cli": "^7.2.9",
+    "@angular/language-service": "^7.2.9",
     "@types/chance": "^1.0.1",
-    "@types/jasmine": "~2.8.3",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~10.3.1",
-    "chance": "^1.0.16",
-    "codelyzer": "^4.0.1",
+    "@types/jasmine": "^2.8.16",
+    "@types/jasminewd2": "^2.0.6",
+    "@types/node": "^10.3.6",
+    "chance": "^1.0.18",
+    "codelyzer": "^4.5.0",
     "jasmine-core": "~3.1.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "~4.0.1",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^2.0.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^1.1.0",
+    "karma-coverage-istanbul-reporter": "^2.0.5",
+    "karma-jasmine": "^1.1.2",
+    "karma-jasmine-html-reporter": "^1.4.0",
     "karma-phantomjs-launcher": "^1.0.4",
-    "mongodb-stitch": "^3.0.0",
-    "protractor": "~5.3.2",
+    "mongodb-stitch": "^3.2.16",
+    "protractor": "^5.4.2",
     "sw-precache-webpack-plugin": "^0.11.5",
-    "ts-node": "~6.1.0",
+    "ts-node": "^8.0.3",
     "tslint": "~5.10.0",
-    "typescript": "^2.7.2",
-    "webpack": "^4.12.0"
+    "typescript": "^3.2.4",
+    "webpack": "^4.29.6"
   }
 }
