@@ -2,6 +2,7 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Contacts } from '@wh-objects/generalData';
 import { Observable } from 'rxjs';
+import { map } from 'rxjs/operators';
 
 @Injectable()
 export class GeneralDataService {
@@ -10,9 +11,11 @@ export class GeneralDataService {
 
   getContacts(): Observable<Contacts> {
     return this.http.get<Contacts>('./assets/content/clubContacts.json')
-      .map(contacts => {
-        console.log(contacts);
-        return contacts;
-      });
+      .pipe(
+        map(contacts => {
+          console.log(contacts);
+          return contacts;
+        })
+      );
   }
 }
