@@ -6,6 +6,7 @@ import { Team, Player } from '@wh-objects/team';
 import { Post } from '@wh-objects/post';
 import { NewsService } from '@wh-share/news.service';
 import { Observable } from 'rxjs';
+import { map } from 'rxjs/operators';
 
 const defaultImg = 'https://wp.willsbach-handball.de/wp-content/uploads/players/avatar_1522109382.png';
 
@@ -21,50 +22,54 @@ export class TeamService {
 
   getPlayer(teamId: string, playerName: string): Observable<Player> {
     return this.http.get<Team>(this.getTeamUrl(teamId))
-      .map(team => {
-        const playerNames = playerName.split('_');
+      .pipe(
+        map(team => {
+          const playerNames = playerName.split('_');
 
-        const player = team.players.find(item =>
-          item.name === playerNames[1] &&
-          item.prename === playerNames[0]
-        );
-        if (player.picture === '' || player.picture === undefined) {
-          // Default Picture if no picture is set
-          player.picture = defaultImg;
-        }
-        return player;
-      });
+          const player = team.players.find(item =>
+            item.name === playerNames[1] &&
+            item.prename === playerNames[0]
+          );
+          if (player.picture === '' || player.picture === undefined) {
+            // Default Picture if no picture is set
+            player.picture = defaultImg;
+          }
+          return player;
+        })
+      );
   }
 
   getTeam(teamId: string): Observable<Team> {
     return this.http.get<Team>(this.getTeamUrl(teamId))
-      .map(team => {
-        // team = team[teamId];
+      .pipe(
+        map(team => {
+          // team = team[teamId];
 
-        if (team.players !== undefined) {
-          team.players.sort(function (a, b) {
-            if (a.prename < b.prename) { return -1; }
-            if (a.prename > b.prename) { return 1; }
-            return 0;
-          });
-        }
-        let picture;
-        this.carousel.getTeam(team.wpID).subscribe(pic => picture = pic,
-          error => { console.log(error); },
-          () => {
-            team.picture = picture[0].media_details.sizes.medium_large.source_url;
-          });
+          if (team.players !== undefined) {
+            team.players.sort(function (a, b) {
+              if (a.prename < b.prename) { return -1; }
+              if (a.prename > b.prename) { return 1; }
+              return 0;
+            });
+          }
+          let picture;
+          this.carousel.getTeam(team.wpID).subscribe(pic => picture = pic,
+            error => { console.log(error); },
+            () => {
+              team.picture = picture[0].media_details.sizes.medium_large.source_url;
+            });
 
-        if (team.trainer !== undefined) {
-          team.trainer.forEach(function (part, index, coach) {
-            if (coach[index].picture === undefined || coach[index].picture === '') {
-              coach[index].picture = defaultImg;
-            }
-          });
-        }
-        this.wpCategory = team.wpCat;
-        return team;
-      });
+          if (team.trainer !== undefined) {
+            team.trainer.forEach(function (part, index, coach) {
+              if (coach[index].picture === undefined || coach[index].picture === '') {
+                coach[index].picture = defaultImg;
+              }
+            });
+          }
+          this.wpCategory = team.wpCat;
+          return team;
+        })
+      );
   }
 
   private getTeamUrl(teamId: string): string {
